requirements.txt
python-telegram-bot>=20.0
apscheduler
sqlalchemy
flask
flask-admin
flask-login
matplotlib
werkzeug

